+++ tencentcloud/data_source_tc_security_groups_test.go
@@ -16,9 +16,12 @@ func TestAccDataSourceTencentCloudSecurityGroups_basic(t *testing.T) {
 				Config: TestAccDataSourceTencentCloudSecurityGroupsConfig,
 				Check: resource.ComposeTestCheckFunc(
 					testAccCheckTencentCloudDataSourceID("data.tencentcloud_security_groups.foo"),
+					resource.TestCheckResourceAttr("data.tencentcloud_security_groups.foo", "security_groups.#", "1"),
 					resource.TestCheckResourceAttr("data.tencentcloud_security_groups.foo", "security_groups.0.name", "ci-temp-security-groups-test"),
 					resource.TestCheckResourceAttr("data.tencentcloud_security_groups.foo", "security_groups.0.description", "ci-temp-security-groups-test"),
 					resource.TestCheckResourceAttr("data.tencentcloud_security_groups.foo", "security_groups.0.be_associate_count", "0"),
+					resource.TestCheckResourceAttr("data.tencentcloud_security_groups.foo", "security_groups.0.ingress.#", "0"),
+					resource.TestCheckResourceAttr("data.tencentcloud_security_groups.foo", "security_groups.0.egress.#", "0"),
 				),
 			},
 		},
@@ -59,6 +62,23 @@ func TestAccDataSourceTencentCloudSecurityGroups_emptyResult(t *testing.T) {
 	})
 }
 
+func TestAccDataSourceTencentCloudSecurityGroups_tags(t *testing.T) {
+	resource.Test(t, resource.TestCase{
+		PreCheck:  func() { testAccPreCheck(t) },
+		Providers: testAccProviders,
+		Steps: []resource.TestStep{
+			{
+				Config: TestAccDataSourceTencentCloudSecurityGroupsTags,
+				Check: resource.ComposeTestCheckFunc(
+					testAccCheckTencentCloudDataSourceID("data.tencentcloud_security_groups.foo"),
+					resource.TestMatchResourceAttr("data.tencentcloud_security_groups.foo", "security_groups.#", regexp.MustCompile(`^[1-9]\d*$`)),
+					resource.TestCheckResourceAttr("data.tencentcloud_security_groups.foo", "security_groups.0.tags.test", "test"),
+				),
+			},
+		},
+	})
+}
+
 const TestAccDataSourceTencentCloudSecurityGroupsConfig = `
 resource "tencentcloud_security_group" "foo" {
   name        = "ci-temp-security-groups-test"
@@ -88,5 +108,20 @@ data "tencentcloud_security_groups" "foo" {
 
 data "tencentcloud_security_groups" "bar" {
   security_group_id = "sg-00000000"
+}
+`
+
+const TestAccDataSourceTencentCloudSecurityGroupsTags = `
+resource "tencentcloud_security_group" "foo" {
+  name        = "ci-temp-security-groups-test"
+  description = "ci-temp-security-groups-test"
+
+  tags = {
+    "test" = "test"
+  }
+}
+
+data "tencentcloud_security_groups" "foo" {
+  tags = "${tencentcloud_security_group.foo.tags}"
 }
 `