+++ tencentcloud/provider.go
@@ -163,12 +163,12 @@ VPC Resources
   tencentcloud_subnet
   tencentcloud_security_group
   tencentcloud_security_group_rule
+  tencentcloud_security_group_lite_rule
   tencentcloud_route_table
   tencentcloud_route_entry
   tencentcloud_route_table_entry
   tencentcloud_dnat
   tencentcloud_nat_gateway
-
 
 */
 package tencentcloud
@@ -330,6 +330,7 @@ func Provider() *schema.Provider {
 			"tencentcloud_gaap_security_policy":       resourceTencentCloudGaapSecurityPolicy(),
 			"tencentcloud_gaap_security_rule":         resourceTencentCloudGaapSecurityRule(),
 			"tencentcloud_ssl_certificate":            resourceTencentCloudSslCertificate(),
+			"tencentcloud_security_group_lite_rule":   resourceTencentCloudSecurityGroupLiteRule(),
 		},
 
 		ConfigureFunc: providerConfigure,